requirements_dev.txt
-r requirements_common.txt

# Fabric
ecdsa==0.13
paramiko==1.17.2
pycrypto==2.6.1
six==1.10.0
Fabric3==1.11.1.post1
